add optional message batching
Normally, messages are send individually. But when there are a lot of
messages queued up that can cause multiple calls to matrix-commander.
This seems to cause matrix.org to deny authentication.

This change adds optional message batching. When enabled, messages are
concatenated into a single file and sent in a single Matrix message.